Add course certificates on the learner profile
LEARNER-1860

Adding a certificates fragment to the learner profile page that shows
the certificates that a user has acquired or a well alerting them
that they have not yet earned certificates and a link to explore new
courses.

Ensures no-certificate well only appears on ones own profile and that
the fragment does not appear when in a limited profile state.

Jasmine and python unit tests added to ensure the above functionality.